[B] refactor: 커스텀 이벤트 객체 이름 수정
