+++ backend/src/main/java/ssu/groupstudy/domain/notification/domain/event/AllUserTopicSubscribeEvent.java
@@ -1,4 +1,4 @@
-package ssu.groupstudy.domain.notification.domain;
+package ssu.groupstudy.domain.notification.domain.event;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,6 +6,6 @@ import ssu.groupstudy.domain.user.domain.User;
 
 @Getter
 @RequiredArgsConstructor
-public class SignInSuccessEvent {
+public class AllUserTopicSubscribeEvent {
     private final User user;
 }
