+++ backend/src/main/java/ssu/groupstudy/domain/notification/listener/SubscribeListener.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import ssu.groupstudy.domain.notification.domain.SignInSuccessEvent;
+import ssu.groupstudy.domain.notification.domain.event.AllUserTopicSubscribeEvent;
 import ssu.groupstudy.domain.notification.domain.TopicCode;
 import ssu.groupstudy.domain.user.domain.User;
 import ssu.groupstudy.global.util.FcmUtils;
@@ -18,7 +18,7 @@ public class SubscribeListener {
     private final FcmUtils fcmUtils;
 
     @EventListener
-    public void handleSignInSuccessEvent(SignInSuccessEvent event) {
+    public void handleSignInSuccessEvent(AllUserTopicSubscribeEvent event) {
         User user = event.getUser();
         fcmUtils.subscribeTopicFor(user.getFcmTokenList(), TopicCode.ALL_USERS, null);
         log.info("## handleSignInSuccessEvent : ");
